benchmark use float type

+++ cpy_opt.h
@@ -2,15 +2,15 @@
 #define CPY_OPT_H
 
 /* load test (with summing to be fair against non_optimized code */
-float cpy_lsopt(double* a,int l,int mx,int it);
+float cpy_lsopt(float *a, int l, int mx, int it);
 
 /* const store test */
-int cpy_vsopt(double* a,int l,int mx,int it);
+float cpy_vsopt(float *a, int l, int mx, int it);
 
 /* load copy test (strided load, contiguous store) */
-int cpy_lcopt(double* a,double* c,int l,int mx,int it);
+float cpy_lcopt(float *a, float *c, int l, int mx, int it);
 
 /* copy store test (contiguous load, strided store) */
-int cpy_csopt(double* a,double* c,int l,int mx,int it);
+float cpy_csopt(float *a, float *c, int l, int mx, int it);
 
 #endif
